conf/layer.conf: Fix SSTATE_EXCLUDEDEPS_SYSROOT
Omit most of the dependencies of rpm-sequoia-native, not all.

The list of dependencies to omit was determined by:

  $ bitbake -e rpm-sequoia-native 2>&1 | grep ^DEPENDS=
  DEPENDS="pkgconfig-native cargo-native openssl-native python3-native rust-native rust-native"

Most notably, openssl-native is needed, but nothing beyond that.
Fortunately, openssl-native was pulled into the native sysroot
of other recipes in via something else, so packaging didn't fail.

But rather be safe than sorry.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>